Bug 6113 [ENH] - enhancement to keep previous ids
Modified the C4::Biblio::_koha_add_biblioitem() && _koha_add_biblio() to support INSERTing biblionumbers and biblioitemnumbers to
the Koha DB. Created validators to make sure the primary keys are positive integers and well below the Perl's LONG_MAX threshold
to keep Koha from getting crazy. Duplicate id's are the users responsibility and in normal use conditions they are non-existing,
unless the "Koha to MARC mapping" to biblionumber is a user-modifiable MARC field, which is TROUBLE!
MySQL takes care of the primary key sequence refreshing.

Updated the bulkmarcimport.pl -script to copy the desired legacy id MARC field to the biblionumber- and biblioitemnumber-fields
designated in "Koha to MARC mapping", eg. copy field 001 to 999c and 999d.
This is handy if you have the legacy id in just 001 and need to populate 999c and 999d to push them to DB.
No need to make an extension to the existing migration scripts.

Documented the new functionality in bulkmarcimport.pl

Created test cases for C4::Biblio.pm modifications.